requirements.txt
openai-whisper
google-genai
pyttsx3
torch
pyserial
watchdog
